fix: moving the accounts to model
